+++ core/src/index/unittest/test_sptag.cpp
@@ -95,6 +95,16 @@ TEST_P(SPTAGTest, sptag_basic) {
         std::cout << "id\n" << ss_id.str() << std::endl;
         std::cout << "dist\n" << ss_dist.str() << std::endl;
     }
+
+    // Though these functions do nothing, use them to improve code coverage
+    {
+        index_->Seal();
+        knowhere::CPUSPTAGRNGIndexModel index_model;
+        // Function Serialize's implementation do'nt have return value,
+        // which will cause undefined behavior.
+        // index_model.Serialize();
+        index_model.Load(knowhere::BinarySet());
+    }
 }
 
 TEST_P(SPTAGTest, sptag_serialize) {