+++ core/src/scheduler/task/BuildIndexTask.cpp
@@ -23,6 +23,7 @@
 #include "utils/Log.h"
 #include "utils/TimeRecorder.h"
 
+#include <fiu-local.h>
 #include <memory>
 #include <string>
 #include <thread>
@@ -60,12 +61,13 @@ XBuildIndexTask::Load(milvus::scheduler::LoadType type, uint8_t device_id) {
                 error_msg = "Wrong load type";
                 stat = Status(SERVER_UNEXPECTED_ERROR, error_msg);
             }
+            fiu_do_on("XBuildIndexTask.Load.throw_std_exception", throw std::exception());
         } catch (std::exception& ex) {
             // typical error: out of disk space or permition denied
             error_msg = "Failed to load to_index file: " + std::string(ex.what());
             stat = Status(SERVER_UNEXPECTED_ERROR, error_msg);
         }
-
+        fiu_do_on("XBuildIndexTask.Load.out_of_memory", stat = Status(SERVER_UNEXPECTED_ERROR, "out of memory"));
         if (!stat.ok()) {
             Status s;
             if (stat.ToString().find("out of memory") != std::string::npos) {
@@ -117,7 +119,8 @@ XBuildIndexTask::Execute() {
         table_file.file_type_ = engine::meta::TableFileSchema::NEW_INDEX;
 
         engine::meta::MetaPtr meta_ptr = build_index_job->meta();
-        Status status = build_index_job->meta()->CreateTableFile(table_file);
+        Status status = meta_ptr->CreateTableFile(table_file);
+        fiu_do_on("XBuildIndexTask.Execute.create_table_success", status = Status::OK());
         if (!status.ok()) {
             ENGINE_LOG_ERROR << "Failed to create table file: " << status.ToString();
             build_index_job->BuildIndexDone(to_index_id_);
@@ -130,6 +133,7 @@ XBuildIndexTask::Execute() {
         try {
             ENGINE_LOG_DEBUG << "Begin build index for file:" + table_file.location_;
             index = to_index_engine_->BuildIndex(table_file.location_, (EngineType)table_file.engine_type_);
+            fiu_do_on("XBuildIndexTask.Execute.build_index_fail", index = nullptr);
             if (index == nullptr) {
                 throw Exception(DB_ERROR, "index NULL");
             }
@@ -150,6 +154,8 @@ XBuildIndexTask::Execute() {
         // step 4: if table has been deleted, dont save index file
         bool has_table = false;
         meta_ptr->HasTable(file_->table_id_, has_table);
+        fiu_do_on("XBuildIndexTask.Execute.has_table", has_table = true);
+
         if (!has_table) {
             meta_ptr->DeleteTableFiles(file_->table_id_);
 
@@ -161,6 +167,7 @@ XBuildIndexTask::Execute() {
 
         // step 5: save index file
         try {
+            fiu_do_on("XBuildIndexTask.Execute.throw_std_exception", throw std::exception());
             status = index->Serialize();
             if (!status.ok()) {
                 ENGINE_LOG_ERROR << status.message();
@@ -171,6 +178,7 @@ XBuildIndexTask::Execute() {
             status = Status(DB_ERROR, msg);
         }
 
+        fiu_do_on("XBuildIndexTask.Execute.save_index_file_success", status = Status::OK());
         if (!status.ok()) {
             // if failed to serialize index file to disk
             // typical error: out of disk space, out of memory or permition denied
@@ -201,6 +209,7 @@ XBuildIndexTask::Execute() {
             status = meta_ptr->UpdateTableFiles(update_files);
         }
 
+        fiu_do_on("XBuildIndexTask.Execute.update_table_file_fail", status = Status(SERVER_UNEXPECTED_ERROR, ""));
         if (status.ok()) {
             ENGINE_LOG_DEBUG << "New index file " << table_file.file_id_ << " of size " << index->PhysicalSize()
                              << " bytes"