+++ core/src/server/delivery/request/CreatePartitionRequest.cpp
@@ -21,6 +21,7 @@
 #include "utils/TimeRecorder.h"
 #include "utils/ValidationUtil.h"
 
+#include <fiu-local.h>
 #include <memory>
 #include <string>
 
@@ -47,22 +48,32 @@ CreatePartitionRequest::OnExecute() {
     try {
         // step 1: check arguments
         auto status = ValidationUtil::ValidateTableName(table_name_);
+        fiu_do_on("CreatePartitionRequest.OnExecute.invalid_table_name",
+                  status = Status(milvus::SERVER_UNEXPECTED_ERROR, ""));
         if (!status.ok()) {
             return status;
         }
 
         status = ValidationUtil::ValidatePartitionName(partition_name_);
+        fiu_do_on("CreatePartitionRequest.OnExecute.invalid_partition_name",
+                  status = Status(milvus::SERVER_UNEXPECTED_ERROR, ""));
         if (!status.ok()) {
             return status;
         }
 
         status = ValidationUtil::ValidatePartitionTags({tag_});
+        fiu_do_on("CreatePartitionRequest.OnExecute.invalid_partition_tags",
+                  status = Status(milvus::SERVER_UNEXPECTED_ERROR, ""));
         if (!status.ok()) {
             return status;
         }
 
         // step 2: create partition
         status = DBWrapper::DB()->CreatePartition(table_name_, partition_name_, tag_);
+        fiu_do_on("CreatePartitionRequest.OnExecute.db_already_exist", status = Status(milvus::DB_ALREADY_EXIST, ""));
+        fiu_do_on("CreatePartitionRequest.OnExecute.create_partition_fail",
+                  status = Status(milvus::SERVER_UNEXPECTED_ERROR, ""));
+        fiu_do_on("CreatePartitionRequest.OnExecute.throw_std_exception", throw std::exception());
         if (!status.ok()) {
             // partition could exist
             if (status.code() == DB_ALREADY_EXIST) {