+++ core/src/utils/ThreadPool.h
@@ -17,6 +17,7 @@
 
 #pragma once
 
+#include <fiu-local.h>
 #include <condition_variable>
 #include <functional>
 #include <future>
@@ -89,7 +90,7 @@ ThreadPool::enqueue(F&& f, Args&&... args) -> std::future<typename std::result_o
 
     auto task = std::make_shared<std::packaged_task<return_type()> >(
         std::bind(std::forward<F>(f), std::forward<Args>(args)...));
-
+    fiu_do_on("ThreadPool.enqueue.stop_is_true", stop = true);
     std::future<return_type> res = task->get_future();
     {
         std::unique_lock<std::mutex> lock(queue_mutex_);