+++ core/unittest/db/test_misc.cpp
@@ -28,6 +28,9 @@
 #include <boost/filesystem.hpp>
 #include <thread>
 #include <vector>
+#include <fiu-local.h>
+#include <fiu-control.h>
+#include "db/utils.h"
 
 TEST(DBMiscTest, EXCEPTION_TEST) {
     milvus::Exception ex1(100, "error");
@@ -91,7 +94,20 @@ TEST(DBMiscTest, UTILS_TEST) {
     options.slave_paths_.push_back("/tmp/milvus_test/slave_2");
 
     const std::string TABLE_NAME = "test_tbl";
-    auto status = milvus::engine::utils::CreateTablePath(options, TABLE_NAME);
+
+    fiu_init(0);
+    milvus::Status status;
+    FIU_ENABLE_FIU("CommonUtil.CreateDirectory.create_parent_fail");
+    status = milvus::engine::utils::CreateTablePath(options, TABLE_NAME);
+    ASSERT_FALSE(status.ok());
+    fiu_disable("CommonUtil.CreateDirectory.create_parent_fail");
+
+    FIU_ENABLE_FIU("CreateTablePath.creat_slave_path");
+    status = milvus::engine::utils::CreateTablePath(options, TABLE_NAME);
+    ASSERT_FALSE(status.ok());
+    fiu_disable("CreateTablePath.creat_slave_path");
+
+    status = milvus::engine::utils::CreateTablePath(options, TABLE_NAME);
     ASSERT_TRUE(status.ok());
     ASSERT_TRUE(boost::filesystem::exists(options.path_));
     for (auto& path : options.slave_paths_) {
@@ -117,6 +133,30 @@ TEST(DBMiscTest, UTILS_TEST) {
 
     status = milvus::engine::utils::DeleteTablePath(options, TABLE_NAME);
     ASSERT_TRUE(status.ok());
+
+    status = milvus::engine::utils::DeleteTableFilePath(options, file);
+    ASSERT_TRUE(status.ok());
+
+    status = milvus::engine::utils::CreateTableFilePath(options, file);
+    ASSERT_TRUE(status.ok());
+
+    FIU_ENABLE_FIU("CreateTableFilePath.fail_create");
+    status = milvus::engine::utils::CreateTableFilePath(options, file);
+    ASSERT_FALSE(status.ok());
+    fiu_disable("CreateTableFilePath.fail_create");
+
+    status = milvus::engine::utils::GetTableFilePath(options, file);
+    ASSERT_FALSE(file.location_.empty());
+
+    FIU_ENABLE_FIU("CommonUtil.CreateDirectory.create_parent_fail");
+    status = milvus::engine::utils::GetTableFilePath(options, file);
+    ASSERT_FALSE(file.location_.empty());
+    fiu_disable("CommonUtil.CreateDirectory.create_parent_fail");
+
+    FIU_ENABLE_FIU("GetTableFilePath.enable_s3");
+    status = milvus::engine::utils::GetTableFilePath(options, file);
+    ASSERT_FALSE(file.location_.empty());
+    fiu_disable("GetTableFilePath.enable_s3");
 
     status = milvus::engine::utils::DeleteTableFilePath(options, file);
     ASSERT_TRUE(status.ok());