+++ core/unittest/metrics/test_prometheus.cpp
@@ -20,8 +20,11 @@
 
 #include <gtest/gtest.h>
 #include <iostream>
+#include <fiu-control.h>
+#include <fiu-local.h>
 
 TEST(PrometheusTest, PROMETHEUS_TEST) {
+    fiu_init(0);
     milvus::server::Config::GetInstance().SetMetricConfigEnableMonitor("on");
 
     milvus::server::PrometheusMetrics instance = milvus::server::PrometheusMetrics::GetInstance();
@@ -57,9 +60,20 @@ TEST(PrometheusTest, PROMETHEUS_TEST) {
     instance.QueryVectorResponseSummaryObserve(1.0, 1);
     instance.QueryVectorResponsePerSecondGaugeSet(1.0);
     instance.CPUUsagePercentSet();
+    fiu_enable("SystemInfo.CPUPercent.mock", 1, nullptr, 0);
+    instance.CPUUsagePercentSet();
+    fiu_disable("SystemInfo.CPUPercent.mock");
     instance.RAMUsagePercentSet();
+    fiu_enable("SystemInfo.MemoryPercent.mock", 1, nullptr, 0);
+    instance.RAMUsagePercentSet();
+    fiu_disable("SystemInfo.MemoryPercent.mock");
     instance.QueryResponsePerSecondGaugeSet(1.0);
     instance.GPUPercentGaugeSet();
+    fiu_enable("SystemInfo.GPUMemoryTotal.mock", 1, nullptr, 0);
+    fiu_enable("SystemInfo.GPUMemoryUsed.mock", 1, nullptr, 0);
+    instance.GPUPercentGaugeSet();
+    fiu_disable("SystemInfo.GPUMemoryTotal.mock");
+    fiu_disable("SystemInfo.GPUMemoryUsed.mock");
     instance.GPUMemoryUsageGaugeSet();
     instance.AddVectorsPerSecondGaugeSet(1, 1, 1);
     instance.QueryIndexTypePerSecondSet("IVF", 1.0);
@@ -71,8 +85,23 @@ TEST(PrometheusTest, PROMETHEUS_TEST) {
     instance.OctetsSet();
 
     instance.CPUCoreUsagePercentSet();
+    fiu_enable("SystemInfo.getTotalCpuTime.open_proc", 1, nullptr, 0);
+    instance.CPUCoreUsagePercentSet();
+    fiu_disable("SystemInfo.getTotalCpuTime.open_proc");
+    fiu_enable("SystemInfo.getTotalCpuTime.read_proc", 1, nullptr, 0);
+    instance.CPUCoreUsagePercentSet();
+    fiu_disable("SystemInfo.getTotalCpuTime.read_proc");
     instance.GPUTemperature();
+    fiu_enable("SystemInfo.GPUTemperature.mock", 1, nullptr, 0);
+    instance.GPUTemperature();
+    fiu_disable("SystemInfo.GPUTemperature.mock");
     instance.CPUTemperature();
+    fiu_enable("SystemInfo.CPUTemperature.opendir", 1, nullptr, 0);
+    instance.CPUTemperature();
+    fiu_disable("SystemInfo.CPUTemperature.opendir");
+    fiu_enable("SystemInfo.CPUTemperature.openfile", 1, nullptr, 0);
+    instance.CPUTemperature();
+    fiu_disable("SystemInfo.CPUTemperature.openfile");
 
     milvus::server::Config::GetInstance().SetMetricConfigEnableMonitor("off");
     instance.Init();