+++ core/unittest/scheduler/test_resource.cpp
@@ -132,13 +132,17 @@ class ResourceAdvanceTest : public testing::Test {
     void
     WaitLoader(uint64_t count) {
         std::unique_lock<std::mutex> lock(load_mutex_);
-        cv_.wait(lock, [&] { return load_count_ == count; });
+        cv_.wait(lock, [&] {
+            return load_count_ == count;
+        });
     }
 
     void
     WaitExecutor(uint64_t count) {
         std::unique_lock<std::mutex> lock(exec_mutex_);
-        cv_.wait(lock, [&] { return exec_count_ == count; });
+        cv_.wait(lock, [&] {
+            return exec_count_ == count;
+        });
     }
 
     ResourcePtr disk_resource_;
@@ -207,6 +211,12 @@ TEST_F(ResourceAdvanceTest, CPU_RESOURCE_TEST) {
     for (uint64_t i = 0; i < NUM; ++i) {
         ASSERT_EQ(tasks[i]->exec_count_, 1);
     }
+
+    std::stringstream out;
+    out << *std::static_pointer_cast<CpuResource>(cpu_resource_);
+    out << *std::static_pointer_cast<GpuResource>(gpu_resource_);
+    out << *std::static_pointer_cast<DiskResource>(disk_resource_);
+    out << *std::static_pointer_cast<TestResource>(test_resource_);
 }
 
 TEST_F(ResourceAdvanceTest, GPU_RESOURCE_TEST) {
@@ -263,6 +273,20 @@ TEST_F(ResourceAdvanceTest, TEST_RESOURCE_TEST) {
     for (uint64_t i = 0; i < NUM; ++i) {
         ASSERT_EQ(tasks[i]->exec_count_, 1);
     }
+
+    test_resource_->TaskAvgCost();
+    std::cout << test_resource_->Dump() << "  " << test_resource_->NumOfTaskToExec() << std::endl;
+    auto null_resource = ResourceFactory::Create("invalid", "invalid", 0);
+    ASSERT_EQ(null_resource, nullptr);
+}
+
+TEST(Connection_Test, CONNECTION_TEST) {
+    std::string connection_name = "cpu";
+    uint64_t speed = 982;
+    Connection connection(connection_name, speed);
+    ASSERT_EQ(connection_name, connection.name());
+    ASSERT_EQ(speed, connection.speed());
+    std::cout << connection.Dump() << std::endl;
 }
 
 }  // namespace scheduler