+++ core/unittest/server/test_cache.cpp
@@ -16,10 +16,13 @@
 // under the License.
 
 #include <gtest/gtest.h>
-#include "cache/CpuCacheMgr.h"
-#include "cache/GpuCacheMgr.h"
+#include <fiu-control.h>
+#include <fiu-local.h>
 #include "utils/Error.h"
 #include "wrapper/VecIndex.h"
+
+#include "cache/CpuCacheMgr.h"
+#include "cache/GpuCacheMgr.h"
 
 namespace {
 
@@ -140,13 +143,17 @@ TEST(CacheTest, CPU_CACHE_TEST) {
     ASSERT_EQ(cpu_mgr->CacheCapacity(), cap);
 
     uint64_t item_count = 20;
-    for (uint64_t i = 0; i < item_count; i++) {
+    for (uint64_t i = 0; i < item_count + 1; i++) {
         // each vector is 1k byte, total size less than 1G
         milvus::engine::VecIndexPtr mock_index = std::make_shared<MockVecIndex>(256, 1000000);
         milvus::cache::DataObjPtr data_obj = std::static_pointer_cast<milvus::cache::DataObj>(mock_index);
         cpu_mgr->InsertItem("index_" + std::to_string(i), data_obj);
     }
     ASSERT_LT(cpu_mgr->ItemCount(), g_num);
+
+    //insert null data
+    std::shared_ptr<milvus::cache::DataObj> null_data_obj = nullptr;
+    cpu_mgr->InsertItem("index_null", null_data_obj);
 
     auto obj = cpu_mgr->GetIndex("index_0");
     ASSERT_TRUE(obj == nullptr);
@@ -170,9 +177,20 @@ TEST(CacheTest, CPU_CACHE_TEST) {
         milvus::cache::DataObjPtr data_obj = std::static_pointer_cast<milvus::cache::DataObj>(mock_index);
         cpu_mgr->InsertItem("index_6g", data_obj);
         ASSERT_TRUE(cpu_mgr->ItemExists("index_6g"));
+
+        //insert aleady existed key
+        cpu_mgr->InsertItem("index_6g", data_obj);
     }
 
     cpu_mgr->PrintInfo();
+
+    cpu_mgr->ClearCache();
+    ASSERT_EQ(cpu_mgr->ItemCount(), 0);
+
+//    fiu_enable("CpuCacheMgr_CpuCacheMgr_ZeroCpucacheThreshold", 1, nullptr, 0);
+//    auto* cpu_cache_mgr = new milvus::cache::CpuCacheMgr();
+//    fiu_disable("CpuCacheMgr_CpuCacheMgr_ZeroCpucacheThreshold");
+//    delete cpu_cache_mgr;
 }
 
 #ifdef MILVUS_GPU_VERSION
@@ -202,6 +220,11 @@ TEST(CacheTest, GPU_CACHE_TEST) {
 
     gpu_mgr->ClearCache();
     ASSERT_EQ(gpu_mgr->ItemCount(), 0);
+
+//    fiu_enable("GpuCacheMgr_GpuCacheMgr_ZeroGpucacheThreshold", 1, nullptr, 0);
+//    auto* gpu_cache_mgr = new milvus::cache::GpuCacheMgr();
+//    fiu_disable("GpuCacheMgr_GpuCacheMgr_ZeroGpucacheThreshold");
+//    delete gpu_cache_mgr;
 }
 #endif
 
@@ -232,4 +255,24 @@ TEST(CacheTest, INVALID_TEST) {
         }
         ASSERT_EQ(mgr.GetItem("index_0"), nullptr);
     }
+}
+
+TEST(CacheTest, PARTIAL_LRU_TEST) {
+    constexpr int MAX_SIZE = 5;
+    milvus::cache::LRU<int, int> lru(MAX_SIZE);
+
+    lru.put(0, 2);
+    lru.put(0, 3);
+    ASSERT_EQ(lru.size(), 1);
+
+    for (int i = 1; i < MAX_SIZE; ++i) {
+        lru.put(i, 0);
+    }
+    ASSERT_EQ(lru.size(), MAX_SIZE);
+
+    lru.put(99, 0);
+    ASSERT_EQ(lru.size(), MAX_SIZE);
+    ASSERT_TRUE(lru.exists(99));
+
+    ASSERT_ANY_THROW(lru.get(-1));
 }