+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Please mark all change in change log and use the issue from GitHub
 - \#1067 - Add binary vectors support in http server
 - \#1075 - improve error message when page size or offset is illegal
 - \#1082 - check page_size or offset value to avoid float
+- \#1115 - http server support load table into memory
 
 ## Feature
 - \#216 - Add CLI to get server info
@@ -33,6 +34,7 @@ Please mark all change in change log and use the issue from GitHub
 - \#813 - Add push mode for prometheus monitor
 - \#815 - Support MinIO storage
 - \#823 - Support binary vector tanimoto/jaccard/hamming metric
+- \#853 - Support HNSW
 - \#910 - Change Milvus c++ standard to c++17
 
 ## Improvement